Run the reverb tank's arithmetic in double, as upstream does
The Dattorro network matched upstream line for line -- same taps, same
diffusion and damping and excursion constants, same tank topology -- but every
intermediate was a float where upstream's is a JavaScript number.  Both keep
their delay lines in float32; the difference is that upstream rounds only when
a sample is stored into one, while the port rounded at every step of a
feedback loop that a sample circulates through for the length of the tail.

Widen the state and the loop: lp[3] and excPhase on the struct, the diffusion
and decay coefficients, the pre-tank and tank intermediates, the excursions and
their trig, the cubic interpolator's taps, and the mix-down.  The delay lines
stay float32 and the write still narrows on store, so what is kept is exactly
what upstream keeps.

The reverb axis was sliding with send level -- 89 dB dry, 77 at CC91=40, 70 at
127 -- which is the signature of a tail that decorrelates as more of it is
audible.  It now reads 94.0/94.4/95.3, flat.  Every other grid axis rises with
it, by 4.5 to 6 dB, because the bank's presets carry a base reverb send and the
grid renders with effects on: the tail was under everything.  The corpus is
unchanged bar gs_filter_test at +1.0 dB.

+++ spessasynth_core/include/spessasynth/synthesizer/dsp/dattorro.h
@@ -40,8 +40,8 @@ typedef struct {
 	float excursionDepth;
 	float gain;
 	float sampleRate;
-	float lp[3];
-	float excPhase;
+	double lp[3];
+	double excPhase;
 	unsigned int pDWrite;
 	unsigned int pDLength;
 	short *taps;
